// InspecWeb/ClientApp/src/app/external-organization/models/Region.ts
// Generated by https://quicktype.io

export interface NewRegion {
    Id:          number;
    Name:        string;
    FiscalYears: FiscalYear[];
}

export interface FiscalYear {
    Year:      number;
    Name:      string;
    Projects:  Projects;
    Provinces: Province[];
}

export interface Projects {
    Count:        number;
    Completed:    number;
    TotalBudget:  number;
    TotalAmount:  number;
    TotalSpent:   number;
    TotalPercent: number;
}

export interface Province {
    Id:   number;
    Name: string;
}